refactored `api_request_handler` to `request_handlers` and seperated graphql request handlers to `graphql_request_handler`
