+++ src/http/mod.rs
@@ -1,7 +1,6 @@
 mod client;
 mod data_loader;
 
-mod api_request_handler;
 mod data_loader_request;
 mod http_1;
 mod http_2;
@@ -9,10 +8,12 @@ mod method;
 pub mod parser;
 mod request_context;
 mod request_handler;
+mod request_handlers;
 mod response;
 mod server;
 mod server_config;
 mod server_context;
+
 use std::time::Duration;
 
 use cache_control::{Cachability, CacheControl};
